Added strikeout for names of finished tasks.
This required creating our own copy of the tabwriter package, in order
to edit how escapes were handled. The only thing changed was the Escape
constant and the endChar for our Escape.

Additionally, changed the default behavior of get to not print finished
tasks

Finally, modified stringify to add finished information.